inventory import/export & shop display fixes
inventory json import and export set up. also a BUNCH of reorganizing because things needed to be in Resources folder, sorry. also separated UI display and function code, and fixed most of the behavior with the shop display.
still TODO:
tweak shopUIDisplayController to get things to look the way we want them to
add the total money to json file for import/export