Fixed shop population/clicking issues
Changed how the buttons interact with opening the shop so that player inventory shows properly. No longer have to double click items to move them around in shop. Added a script called ConsoleToGui that displays the console in a gui panel (for use in debugging builds). CursorDisplayHandler added to player inventory in main scene to supplement the double click fix.